fixed toutatice cas
