+++ src/cas/generics/educonnect.js
@@ -1,8 +1,8 @@
 const errors = require('../../errors');
 const { getDOM, submitForm, extractStart } = require('../api');
+const http = require('../../http');
 
-async function login({ dom, jar, url, account, username, password })
-{
+async function login({ dom, jar, url, account, username, password }) {
     dom.window.document.getElementById('username').value = username;
     dom.window.document.getElementById('password').value = password;
 
@@ -19,13 +19,7 @@ async function login({ dom, jar, url, account, username, password })
         throw errors.WRONG_CREDENTIALS.drop();
     }
 
-    await submitForm({ dom, jar });
-
-    return extractStart(await getDOM({
-        url: url + account.value + '.html',
-        jar,
-        asIs: true
-    }));
+    return await submitForm({ dom, jar, followRedirects: false });
 }
 
 module.exports = login;
